Set add function and context to buttons
Added an empty Float array to hold individual receipt items.

Added a boolean value to enable a start to the label onscreen.

Attached an IBAction to each of the number buttons onscreen.

Created IBAction for the add and subtract buttons.

When the user presses the add button, append the label contents to the
receiptItems array, casting the string value as a Float.