Till screen and receipt screen properly connected
Added the math to process input in order to display a proper Float
value.

Ensured that the items properly appended themselves to an array.

Set up a segue function to pass the receipt information from the till
to the receipt screen.

Designed and properly implemented the receipt table view.

Added the ability for the program to add all of the individual items to
display a subtotal.

Completely created the framework for the Receipt screen.

Started the framework for the Batch screen.