Set up saving Core Data
Set up the capabilities and framework for saving/loading core data.

Created a date and items attribute for the Receipt entity.

Created a date and totals attribute for the Batch entity.

Ensured that the managedContext object would appropriately save to the
device utilizing the .save() parameter in both the Batch VC and the
Receipt VC.