CocoaPods added, final touches added
Added functionality for cocoa pods in order to install the Tesseract
Library.

There is a known error that has no solution in the functionality of
Tesseract which causes the application to be unusable on an actual
device, which is needed for testing of the camera functions. The only
fix for the error currently is to rebuild the entire project.

Added the check view controller, and removed unnecessary View
Controllers.

Removed and added remarks and comments.

Removed unnecessary code.